+++ ASD-support/src/main/java/com/demo/asd/support/model/po/staff/StaffCriteria.java
@@ -8,7 +8,7 @@ import lombok.Setter;
  */
 public class StaffCriteria {
     //员工ID
-    public Integer staffId;
+    public Long staffId;
 
     //员工名
     public String staffUsername;
@@ -20,13 +20,13 @@ public class StaffCriteria {
     public Integer staffLevel;
 
     //员工所在分行号
-    public Integer staffBranchId;
+    public Long staffBranchId;
 
     //员工职位
     public String staffJob;
 
     //签到临时流水号
-    public Integer todaySerialId;
+    public Long todaySerialId;
 
     //员工头像
     public String staffImg;
@@ -34,11 +34,11 @@ public class StaffCriteria {
     //1启用,0注销
     public Integer status;
 
-    public Integer getStaffId() {
+    public Long getStaffId() {
         return staffId;
     }
 
-    public void setStaffId(Integer staffId) {
+    public void setStaffId(Long staffId) {
         this.staffId = staffId;
     }
 
@@ -66,11 +66,11 @@ public class StaffCriteria {
         this.staffLevel = staffLevel;
     }
 
-    public Integer getStaffBranchId() {
+    public Long getStaffBranchId() {
         return staffBranchId;
     }
 
-    public void setStaffBranchId(Integer staffBranchId) {
+    public void setStaffBranchId(Long staffBranchId) {
         this.staffBranchId = staffBranchId;
     }
 
@@ -82,11 +82,11 @@ public class StaffCriteria {
         this.staffJob = staffJob;
     }
 
-    public Integer getTodaySerialId() {
+    public Long getTodaySerialId() {
         return todaySerialId;
     }
 
-    public void setTodaySerialId(Integer todaySerialId) {
+    public void setTodaySerialId(Long todaySerialId) {
         this.todaySerialId = todaySerialId;
     }
